added selecting code, DOESN'T DISPLAY
added selecting code.
added unselecting code.

added boolean wonGame

added newBoard() //displays new board

_____
NEED TO: fix flashing code so it doesn't make the entire board flash.

add/figure out code to make it possible for it to tell when the board
is complete and then to make wonGame = true.
idea is to make this in a new method and then add it in under
if(Button_A)

___________
Problems:
The board will not display on the meggy. Not sure why. It compiles and
it isn't the battery because my snake game worked.

Oh, also the flashing part. Need to fix that. It flashes the entire
board. Use code similar to what will be used to tell whether or not it
should be wonGame or not. Perhaps throw them in the same method.