99% Done, WILL NOT DISPLAY
Code is complete.
Added:
- if you press A and B, the board will regenerate (contingency incase
there is an unsolvable board)

- A selects, B unselects points

- You can 'level up'
- - checks to see if there are any same colors on column/row, if there
aren't, plays sound and creates new board

- flashing code should now work.

PROBLEM:
Have not been able to test any of this because, although it compiles,
it will not upload to the meggy. Other games will upload, and I just
put in new batteries maybe 5, 10 minutes ago to see if that would help
(it did, but Unfriendly Colors still does not work). Instead of
displaying anything, it just sounds a dull bzzz.

Once that is fixed, need to:
clean up code. Indenting and everything of that nature is good, but
clean up/out extra comments and broken, commented out code.