debugging.
So, the game displays.
Curser works, for the most part.

Colors are weird.. they follow no pattern and sometimes listen to what
they're supposed to do. I have a growing inclination that it's a
hardware problem.

However, there is something wrong with checkAllColors becaues even if
they're all differen't colors (or appear to be) it doesn't work…

Need to try someone else's meggy because this one is so erratic I
cannot tell whether it is me or it that is wrong.

Also, the colors on the same line do not flash. That's part of the code
that would cause it to then win and regenerate, so I'm assuming that's
all broken. No idea. Must have someone else look at it. Compiles and
looks too good. No idea.